一応現在地からの取得もいけた

+++ app/src/main/java/jp/ac/it_college/std/s22025/weather_api_exam/model/WhetherInfo.kt
@@ -58,7 +58,7 @@ data class MainContents(
     @SerialName("temp_max") val tempMax: Double,
 
     //気圧
-    val pressure: Double,
+    val pressure: Int,
 
     //湿度
     val humidity: Double
